+++ src/components/ClientProviders.tsx
@@ -12,6 +12,7 @@ import { SupabaseAuthProvider } from '@/contexts/SupabaseAuthContext';
 import { SiteConfigProvider } from '@/contexts/SiteConfigContext';
 import { AlertProvider } from '@/contexts/AlertContext';
 import { SignatureProvider } from '@/contexts/SignatureContext';
+import { CompanionSessionProvider } from '@/contexts/CompanionSessionContext';
 // Using our new safer approach for Material Design icons
 import MaterialDesignIcon from '@/components/MaterialDesignIcon';
 import LanguageDialog from '@/components/LanguageDialog';
@@ -74,22 +75,24 @@ export default function ClientProviders({ children }: { children: React.ReactNod
     <>
       <GlobalErrorHandler />
       <SupabaseAuthProvider>
-        <I18nProvider>
-          <SiteConfigProvider>
-            <AlertProvider>
-              <SignatureProvider>
-                <GlobalTimeTracker />
-                <SiteHead />
-                {isMounted && <LanguageDialog />}
-                {isMounted && <ToastContainer position="top-right" theme="colored" />}
-                {isMounted && <Toaster position="top-right" />}
-                {isMounted && <ChangelogModal />}
-                <ProfilePromptGate isMounted={isMounted} pathname={pathname} />
-                {children}
-              </SignatureProvider>
-            </AlertProvider>
-          </SiteConfigProvider>
-        </I18nProvider>
+        <CompanionSessionProvider>
+          <I18nProvider>
+            <SiteConfigProvider>
+              <AlertProvider>
+                <SignatureProvider>
+                  <GlobalTimeTracker />
+                  <SiteHead />
+                  {isMounted && <LanguageDialog />}
+                  {isMounted && <ToastContainer position="top-right" theme="colored" />}
+                  {isMounted && <Toaster position="top-right" />}
+                  {isMounted && <ChangelogModal />}
+                  <ProfilePromptGate isMounted={isMounted} pathname={pathname} />
+                  {children}
+                </SignatureProvider>
+              </AlertProvider>
+            </SiteConfigProvider>
+          </I18nProvider>
+        </CompanionSessionProvider>
       </SupabaseAuthProvider>
     </>
   );